feat: Migrate a majority to python package with toml configurations (#4)
There is a LOT of work in this PR. Essentially, most everything is migrated except for `root2hdf5` and `sherpa` scripts. All the top-level scripts are being removed and will be revisited and migrated over to various configs later.

The python packaging is set up such that all the various data files that are not part of the package are installed as `wheel` data, which is discoverable via `sysconfig`. The `sdist` contains everything to remake the wheel as needed. For now, this is a universal wheel as we have no extensions to build, and everything relies (right now) on `docker` existing on the system.