feat: clean up docker workflow and fix pyplotting-cuda dockerfile (#10)
This cleans up the docker workflow by only running the generation after
building the various docker images if they need to be built. Additional logic
was added in to allow for dynamically generating the list of docker images to
build based on changed files instead and injecting it into the matrix. It
should be more robust.